# public/openapi.yaml
openapi: "3.0.3"
info:
  title: 智学AI - 开放API
  description: |
    智学AI教育平台开放API接口文档。提供AI辅导、LLM代理、题库管理、游戏化学习等接口。

    ## 快速开始
    1. **注册账号** - 在智学AI平台注册账号
    2. **获取API Key** - 在个人中心的「API Keys」页面创建虚拟API Key
    3. **调用API** - 使用 `X-API-Key` 请求头携带API Key发起请求

    ## 认证方式
    - **API Key** (`ApiKeyAuth`): 通过 `X-API-Key` 请求头或 `Authorization: Bearer <key>` 传递
    - **会话认证** (`BearerAuth`): 通过 NextAuth 的 session cookie 自动认证（仅限平台内调用）

    ## 错误响应
    所有接口在发生错误时都会返回统一格式：
    ```json
    { "error": "错误描述信息" }
    ```

    常见状态码：
    - `400` - 请求参数错误
    - `401` - 未认证 / API Key无效
    - `402` - 余额不足
    - `403` - 无权限
    - `404` - 资源不存在
    - `429` - 请求频率超限
    - `500` - 服务器内部错误

    ## 速率限制
    - 免费用户: 50 RPM
    - VIP用户: 200 RPM
    超出限制后返回 `429` 状态码。

  version: "1.0.0"
  contact:
    name: 智学AI开发团队
    email: dev@zixueai.com
    url: https://zixueai.com
servers:
  - url: "{baseUrl}"
    description: 智学AI API服务器
    variables:
      baseUrl:
        default: "https://api.zixueai.com"
        description: API服务器地址
        enum:
          - "https://api.zixueai.com"
          - "http://localhost:3000"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "从个人中心获取的API Key，格式: sk-edu-xxxxxx-xxxxxxxx"
    BearerAuth:
      type: http
      scheme: bearer
      description: "NextAuth Session Token（平台内自动携带）"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: 错误描述信息
          example: "请先登录"
      required:
        - error
...
    ChatRequest:
      type: object
      properties:
        message:
          type: string
          description: 用户发送的消息内容
          example: "请帮我理解勾股定理"
          minLength: 1
        conversationId:
          type: string
          description: 对话ID（可选，用于继续已有对话）
          example: "conv_abc123"
        subject:
          type: string
          description: 学科
          example: "数学"
        stream:
          type: boolean
          default: false
          description: 是否使用流式响应
      required:
        - message
...
    ChatResponse:
      type: object
      properties:
        id:
          type: string
          example: "msg_abc123"
        conversationId:
          type: string
          example: "conv_abc123"
        role:
          type: string
          example: "assistant"
        content:
          type: string
          example: "勾股定理是直角三角形两条直角边的平方和等于斜边的平方..."
        stage:
          type: string
          example: "guide"
        tokens:
          type: object
          properties:
            input:
              type: integer
              example: 45
            output:
              type: integer
              example: 120
        cost:
          type: number
          example: 0.0003
      required:
        - id
        - role
        - content
...
    BalanceResponse:
      type: object
      properties:
        balance:
          type: number
          description: 账户余额（CNY）
          example: 12.50
        currency:
          type: string
          example: "CNY"
...
    LLMChatRequest:
      type: object
      properties:
        model:
          type: string
          description: 模型名称
          example: "deepseek-chat"
        messages:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [system, user, assistant]
              content:
                type: string
          minItems: 1
        stream:
          type: boolean
          default: false
          description: 是否流式响应
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 0.7
        max_tokens:
          type: integer
          minimum: 1
          maximum: 16384
          default: 2048
      required:
        - model
        - messages
    LLMChatResponse:
      type: object
      properties:
        id:
          type: string
          example: "chatcmpl-abc123"
        object:
          type: string
          example: "chat.completion"
        created:
          type: integer
          example: 1746600000
        model:
          type: string
          example: "deepseek-chat"
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              message:
                type: object
                properties:
                  role:
                    type: string
                    example: "assistant"
                  content:
                    type: string
              finish_reason:
                type: string
                example: "stop"
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
            cost:
              type: number
              description: 本次请求费用（CNY）
...
paths:
...
  /api/ai/chat:
    post:
      tags:
        - AI Tutoring
      summary: AI辅导对话
      description: |
        向AI老师发送消息，获取苏格拉底式教学辅导。支持会话记忆，可通过 `conversationId` 继续已有对话。

        教学流程采用苏格拉底引导式，阶段包括：
        - `diagnostic` - 诊断评估
        - `clarification` - 澄清理解
        - `hypothesis` - 提出假设
        - `reasoning` - 推理论证
        - `reflection` - 反思总结
        - `verification` - 验证掌握
        - `guide` - 引导推进
      operationId: aiChat
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
            examples:
              newConversation:
                summary: 新建对话
                value:
                  message: "请帮我理解勾股定理"
                  subject: "数学"
              continueConversation:
                summary: 继续已有对话
                value:
                  message: "那如何用勾股定理证明两个三角形相似呢？"
                  conversationId: "conv_abc123"
                  subject: "数学"
      responses:
        "200":
          description: AI回复
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
              example:
                id: "msg_xyz789"
                conversationId: "conv_abc123"
                role: "assistant"
                content: "好的，让我们从勾股定理的基本定义开始理解。首先，你能告诉我什么是直角三角形吗？"
                stage: "diagnostic"
                tokens:
                  input: 45
                  output: 98
                cost: 0.0002
        "400":
          description: 参数错误或内容安全问题
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "消息内容不能为空"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /llm-api/chat:
    post:
      tags:
        - LLM Proxy
      summary: LLM聊天补全（OpenAI兼容）
      description: |
        公开LLM代理接口，兼容 OpenAI Chat Completions API 格式。
        使用您的 API Key 进行认证，支持 Provider 健康检查和故障转移。

        **认证方式**: `X-API-Key` 请求头或 `Authorization: Bearer` 请求头

        **支持的Provider**: deepseek, qwen, openai, ollama, custom
      operationId: llmChatCompletions
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LLMChatRequest"
            examples:
              basic:
                summary: 基础聊天
                value:
                  model: "deepseek-chat"
                  messages:
                    - role: "system"
                      content: "你是一位专业的数学老师。"
                    - role: "user"
                      content: "请解释什么是微积分？"
                  temperature: 0.7
                  max_tokens: 2048
              streaming:
                summary: 流式响应
                value:
                  model: "deepseek-chat"
                  messages:
                    - role: "user"
                      content: "写一首关于学习的诗"
                  stream: true
      responses:
        "200":
          description: 聊天补全结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LLMChatResponse"
              example:
                id: "chatcmpl-abc123"
                object: "chat.completion"
                created: 1746600000
                model: "deepseek-chat"
                choices:
                  - index: 0
                    message:
                      role: "assistant"
                      content: "微积分是数学的一个分支，主要研究函数的变化率和累积量..."
                    finish_reason: "stop"
                usage:
                  prompt_tokens: 45
                  completion_tokens: 200
                  total_tokens: 245
                  cost: 0.0006
        "400":
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "模型不能为空"
        "401":
          description: API Key无效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Invalid API key"
        "402":
          description: 余额不足
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "余额不足，请充值后重试"
        "429":
          description: 请求频率超过限制
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "请求频率过高，请稍后重试"
...
  /api/billing/balance:
    get:
      tags:
        - Billing
      summary: 查询账户余额
      description: 获取当前用户的账户余额（优先从Redis读取，降级到数据库）
      operationId: getBalance
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 余额信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceResponse"
              example:
                balance: 12.50
                currency: "CNY"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...